+++ upgrade/v13xto140/upgrade.go
@@ -1,12 +1,9 @@
 package v13xto140
 
 import (
-	"context"
-
 	"github.com/pkg/errors"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
 
 	longhorn "github.com/longhorn/longhorn-manager/k8s/pkg/apis/longhorn/v1beta2"
@@ -20,9 +17,6 @@ const (
 
 func UpgradeResources(namespace string, lhClient *lhclientset.Clientset, kubeClient *clientset.Clientset, resourceMaps map[string]interface{}) error {
 	if err := upgradeVolumes(namespace, lhClient, resourceMaps); err != nil {
-		return err
-	}
-	if err := upgradeWebhookAndRecoveryService(namespace, kubeClient); err != nil {
 		return err
 	}
 	return nil
@@ -50,28 +44,6 @@ func upgradeVolumes(namespace string, lhClient *lhclientset.Clientset, resourceM
 		}
 		if v.Spec.UnmapMarkSnapChainRemoved == "" {
 			v.Spec.UnmapMarkSnapChainRemoved = longhorn.UnmapMarkSnapChainRemovedIgnored
-		}
-	}
-
-	return nil
-}
-
-func upgradeWebhookAndRecoveryService(namespace string, kubeClient *clientset.Clientset) error {
-	selectors := []string{"app=longhorn-conversion-webhook", "app=longhorn-admission-webhook", "app=longhorn-recovery-backend"}
-
-	for _, selector := range selectors {
-		deployments, err := kubeClient.AppsV1().Deployments("").List(context.TODO(), metav1.ListOptions{LabelSelector: selector})
-		if err != nil {
-			if apierrors.IsNotFound(err) {
-				continue
-			}
-			return errors.Wrapf(err, upgradeLogPrefix+"failed to get deployment with label %v during the upgrade", selector)
-		}
-		for _, deployment := range deployments.Items {
-			err := kubeClient.AppsV1().Deployments(deployment.Namespace).Delete(context.TODO(), deployment.Name, metav1.DeleteOptions{})
-			if err != nil {
-				return errors.Wrapf(err, upgradeLogPrefix+"failed to delete the deployment with label %v during the upgrade", selector)
-			}
 		}
 	}
